requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
google-generativeai==0.5.0
google-adk==0.1.0
mcp==0.9.0
asyncpg==0.29.0
pgvector==0.2.4
aiokafka==0.10.0
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.0
locust==2.23.1
structlog==24.1.0
python-multipart==0.0.9
requests==2.31.0
kafka-python==2.0.2
